# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaksi Pekerjaan API",
        "description": "API untuk aplikasi **Transaksi Pekerjaan** — pencatatan arus kas per kegiatan\nbeserta perhitungan bagi hasil.\n\n### Alur autentikasi\n1. `POST /api/auth/login` dengan email + password → dapat `access_token`.\n2. Kirim `Authorization: Bearer <access_token>` untuk semua endpoint lain.\n3. Opsional: `POST /api/auth/biometric/enable` (butuh password lagi) →\n   dapat `biometric_token` yang disimpan aplikasi di Android Keystore.\n   Login berikutnya cukup sidik jari lewat `POST /api/auth/biometric/login`.\n\n### Aturan angka\nSemua nilai uang adalah **rupiah bulat** (integer, tanpa desimal).\nPersentase dikirim sebagai angka desimal biasa, mis. `1.75` untuk 1,75%.",
        "contact": {
            "email": "bayuapriansah10@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/aktivitas/saya": {
            "get": {
                "tags": [
                    "Aktivitas"
                ],
                "summary": "Riwayat aktivitas sendiri",
                "description": "Riwayat aktivitas pengguna yang sedang login. Semua peran boleh mengaksesnya, dan hanya berisi jejaknya sendiri.",
                "operationId": "aktivitasSaya",
                "parameters": [
                    {
                        "name": "modul",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "auth",
                                "kegiatan",
                                "bahan_baku",
                                "kas",
                                "lampiran",
                                "pengguna",
                                "pengaturan",
                                "lain"
                            ]
                        }
                    },
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "hanya_gagal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/aktivitas": {
            "get": {
                "tags": [
                    "Aktivitas"
                ],
                "summary": "Riwayat aktivitas semua akun",
                "description": "Riwayat aktivitas seluruh pengguna. Khusus superadmin. Bisa disaring per pengguna lewat parameter user_id untuk menelusuri pekerjaan satu petugas.",
                "operationId": "aktivitasSemua",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Saring untuk satu pengguna",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "modul",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari pada aksi, subjek, atau nama pengguna",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hanya_gagal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Bukan superadmin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/aktivitas/ringkasan": {
            "get": {
                "tags": [
                    "Aktivitas"
                ],
                "summary": "Ringkasan aktivitas",
                "description": "Angka ringkas untuk layar Log Aktivitas: jumlah aktivitas per modul, per pengguna, dan jumlah yang gagal. Khusus superadmin.",
                "operationId": "aktivitasRingkasan",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan access_token dari endpoint login (tanpa kata \"Bearer\").",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}